Add new window choice picker action
The new action, `select_window_choices`, allows users to select a window
from the current tabpage and then presents a list of choices to perform
on that window.

The idea is this can be combined with existing actions to create more
complex workflows. For example, you could create a choice picker that
lets the user choose between focusing that window, swapping the selected
window with another or resizing the window.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>