Add extend module to replace vim.tbl_extend and vim.tbl_deep_extend usage
vim.tbl_extend and vim.tbl_deep_extend are used in multiple places
throughout the codebase but it doesn't handle objects with metatables
properly. This is primarily an issue in places where actions are used
with mappings, as the actions are objects with metatables, and merging
them using the built-in functions results in the object losing it's
metatable.

This commit adds a new module `viewport.extend` that provides
`extend` and `deep_extend` functions that correctly handle metatables.
The existing code is then updated to use these new functions instead of
the built-in ones.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>